Started working on player draws
The plan is to move cards in the player hand similar to how the allies move when one dies at the end of every turn. This will allow new cards to be added. Started by adding a place to move cards from "Cards in Hand" to "cards that have been played" in PlayerHandler. This will necessitate a small change in the code to move allies after death.