Drawing Cards Continues.
Changed the code around dealing the starting hand to work better with the new "drawCard" function that is being written to draw new cards throughout the game. It's going to be important to track which cards are in hand and which are in play in order to make sure that everything ends up where it needs to be.