Drag+Drop Attacking and Hero
Allies no longer attack randomly at the end of the turn. Players can now drag a played ally onto an enemy to have that ally attack that enemy once per turn. Additionally, the Hero now spawns in the hero zone, and can attack once per turn as well in the same way. Added a HeroHandler for the hero, and moved STARTINGHANDSIZE to the RulesHandler.